requirements.txt
openai
pandas
python-dotenv
faiss-cpu
streamlit
python-decouple
redis